Add ITALIC, CONCEAL, BGBOLD and VT100 blinking.
CSI 2 m     Adjust code for setting screen colour to dim.
            DIM disables BOLD and vs-versa
CSI 3 m     Set screen colour to italic (another reverse video)
CSI 8 m     Hide the character printed by making the normal display
            just the background colour. If the text is selected the
            text is shown
CSI 6 m     With the advent of 256 colours the SCO Unix bright background
            sequence can now be implemented using a distinct flag,
            it doesn't override the 'blink_is_real' flag.

Mixed in with this is a control to switch the merging of attributes
between the ugly PC style and the much better thought out VT100 style.

As described in the VT100 technical manual the VT100 blinking works with
the other attributes; it turns off Reverse or Bold in preference to using
the default of setting the intensity low. In addition the combination of
reverse plus bold is a useful combination which increases the contrast
between foreground and background rather than decreasing it.

Obviously this is for compatibility with VT100 or PCVGA only, new code
should minimise mixing of attributes.