Add new interpretation for XTerm SGR 38/48
After actually seeing a copy of ITU T.416-199303 XTerm was modified
to be able to use the character ':' as the parameter separator for 256
colour mode.

Unfortunatly, this documentation is still very unclear and the section
looks like it may have been written without fully understanding the
confused phrasing of ECMA-48.

However, the exact interpretation is not really relevent as for
compatibility reasons all variations will have to be supported.

As this is the only sequence that officially uses ':', is likely to be
the only one that ever uses that character and XTerm in effect does it
too. I've simply simply made ':' an alias of ';'.

On review it seems foolish to allow users to give us ANY sequence with
semicolons replaced by colons. So I've made this more complex than just
adding the alias and only allow the ':' for the SGR sequence.