Support of 256 colour palettes and broken drivers.
PuTTY will now redraw it's window when the system palette changes. In
addition the default colours are now at the start of the palette given
to Windows so it's more likely to choose a good colour when the user is
most interested.

The old method using 'UpdateColors' would mostly work on some devices, on
others it would give a completely corrupted set of screen colours. Even
when it did work if you were using XTerm 256 colour mode it would
be unlikely that the colours would ever return to the correct visual
representation and the errors could even spoil the simple 16 colour
collection. The new method simply does a complete redraw when the palette
changes; this is much more robust in the face of dodgy drivers and means
that the colours will always be as correct as possible.

As it now seems stable (and works for Windows95) I've turned it on by
default when Windows reports that the display has a physical palette.

Note that the 6x6x6 cube is scrambled in the palette given to Windows.
This means that even if Windows only gives us part of it we should get
a good selection of colours.

In addition I've finally given up on using the hDC returned by BeginPaint.
Instead the BeginPaint/EndPaint pair is used to fetch the invalidated
rectangle and tell Windows that it'll be fixed.

After that we get a proper DC for the entire window and use that for
all the drawing.

This allows us to always update the screen immediatly during the WM_PAINT
message processing and avoid all the bugs in other applications that seem
to trip us when we assume the clip region that BeginPaint sets up is okay.

The end result is that it appears to be much more reliable and means
that the 256 colour mode screen reacts nice and quickly.