Add XTerm extensions for VTTEST and reverse BS
First is a the XTerm code to enable and disable the DECCOLM 132/80
column switching DEC control which stops the the default xterm 'init'
sequence from terminfo clearing the screen.

The second is XTerm's reverse BS hack to 'pass' the vttest option 1.
This means that PuTTY uses the same modal hack as XTerm to pass the vttest
'end of line' test. Actually that part of the test isn't testing the end
of line processing but instead it detects what happens for a backspace
that is logically at the start of the next line, but the cursor movement
is still pending due to the VT102 eol glitch.

Thirdly, according to ECMA-48, it is undefined what happens if the REP
control does not come immediatly after a printable character. XTerm
defines this to do nothing, vttest has added an incorrect test that
invokes this undefined behaviour. This change prints nasal demons
for vttest.

In addition most applications realise that terminal resizes are requests
and that you likely won't get what you ask for.  Programs (like vttest)
that use the sequence DECCOLM aren't so smart so the change is immediatly
enacted.

In summary, these use the XTerm vttest compatibility flag to work as
vttest expects rather than the more generally useful generic compatibility
style.