Readme and TODO for github with other bits
A .gitignore file

A README.md for github

A working GUNmakefile with proper cleanup that builds both the Linux host and
Windows cross compile at the same time.

An old Windows makefile to call nmake 'correctly'

A TODO file with some minor idea notes

Some files with the codes understood by each version of PuTTY.

Fixup for new PuTTY version scheme

Simon isn't supporting git-describe

This patch searches for the tags after 0.63 in a different way.

The main result is that it's now using the shortest route between the
tags and the HEAD to choose which numbers to put on as the version.

It uses git show-branch as this seems to be unique in the way it chooses
which commits to show.

Test compile before push and ...

create a branch with the executables to push on their own.

Find closest rgbCIEDE2000 colour from palette.
    I don't like the results.

Tried to use mksrcarc.sh but it zips up eveything including the .git
directory.  So I threw my own together.

Make macosx directory optional

Add 'fetch' rule to Makefile.rdb