ITU-T T.416 truecolour (Phase 2)
Colours are displayed as full truecolour.

Windows and Unix truecolour screens are working, even a 256 colour
windows display mode gives a reasonable interpretation.

This commit works in four different modes.
    Extended termchar implementation
        32 bit colour fields in termchar; saved to scrollback.
    Thin extended termchar implementation
        32 bit colour fields in termchar; NOT saved to scrollback.
        #define THIN_TC24
    FAT COMBI implementation
        24 bit colours stored in cc_char list, saved to scrollback.
        #define FAT_COMBI_COLOUR
    THIN COMBI implementation
        24 bit colours stored in cc_char list, NOT saved to scrollback.
        #define COMBI_COLOUR

None of these have additional memory requirements for the scrollback
memory unless 24 bit colours are actually used.

If truecolour is used the 'Thin' versions only display it on the main
screen, the scrollback is reduced to just 256 colours.

The fastest implementation is the 'fat' extended termchar one.

If truecolour is used the two thin variants have about the same memory
consumption. The FAT-COMBI variant uses the most memory when truecolour
is used, by a large percentage (it is also the slowest).

When truecolour is not used the memory requirements of the versions differ
by only a few kilobytes, the extended termchar variants should, in theory,
take more memory, but this is unconfirmed. The thin variants are the
slowest in this case, the the fat variants somewhat faster, the "extended
termchar" variant may be fractionally faster with the test cases.

Also, currently included:
    ITU-T T.416 truecolour, Extend the palette sequences for SGR 38

    Colours are displayed as full truecolour.

    Colours understood by SGR 38;5;N and SGR 48;5;N become:

    Range           Size    Use
    0..15           16      Ansi 8 and 16 colour sets
    16..231         216     XTerm Colour cube
    232..255        24      XTerm greys
    256..4351       4096    12 bit cube
    4352..4607      256     Full grey scale

    16777216..33554431 24bit colour with values 2**24 + 65536*r + 256*g + b